pyproject.toml
[tool.poetry]
name = "ai-agent"
version = "0.1.0"
description = "一个遵循Unix哲学的AI命令行代理"
authors = ["Bowen Yang <ybw051114@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1.3"
openai = "^1.12.0"
rich = "^13.7.0"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
tomli-w = "^1.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.1.0"
isort = "^5.12.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ai-agent = "ai_agent.cli:main"
